move Result.kt directory
