+++ app/src/main/java/com/bangkit/alpaca/data/AuthRepository.kt
@@ -2,7 +2,7 @@ package com.bangkit.alpaca.data
 
 import com.bangkit.alpaca.data.remote.firebase.IFirebaseAuth
 import com.bangkit.alpaca.model.User
-import com.bangkit.alpaca.utils.Result
+import com.bangkit.alpaca.data.remote.Result
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.database.FirebaseDatabase
 import kotlinx.coroutines.flow.flow
