+++ app/src/main/java/com/bangkit/alpaca/ui/auth/login/LoginFragment.kt
@@ -11,7 +11,7 @@ import androidx.navigation.findNavController
 import com.bangkit.alpaca.R
 import com.bangkit.alpaca.databinding.FragmentLoginBinding
 import com.bangkit.alpaca.ui.main.MainActivity
-import com.bangkit.alpaca.utils.Result
+import com.bangkit.alpaca.data.remote.Result
 import com.bangkit.alpaca.utils.isTouchableScreen
 import com.bangkit.alpaca.utils.showError
 import com.bangkit.alpaca.utils.showToastMessage
