+++ app/src/main/java/com/bangkit/alpaca/ui/auth/registration/RegistrationViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.bangkit.alpaca.data.AuthRepository
-import com.bangkit.alpaca.utils.Result
+import com.bangkit.alpaca.data.remote.Result
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
